+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Collections/NetworkList.cs
@@ -369,7 +369,8 @@ namespace Unity.Netcode
             // check write permissions
             if (!CanClientWrite(m_NetworkBehaviour.NetworkManager.LocalClientId))
             {
-                throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
+                LogWritePermissionError();
+                return;
             }
 
             m_List.Add(item);
@@ -390,7 +391,8 @@ namespace Unity.Netcode
             // check write permissions
             if (!CanClientWrite(m_NetworkBehaviour.NetworkManager.LocalClientId))
             {
-                throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
+                LogWritePermissionError();
+                return;
             }
 
             m_List.Clear();
@@ -416,7 +418,8 @@ namespace Unity.Netcode
             // check write permissions
             if (!CanClientWrite(m_NetworkBehaviour.NetworkManager.LocalClientId))
             {
-                throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
+                LogWritePermissionError();
+                return false;
             }
 
             int index = m_List.IndexOf(item);
@@ -451,7 +454,8 @@ namespace Unity.Netcode
             // check write permissions
             if (!CanClientWrite(m_NetworkBehaviour.NetworkManager.LocalClientId))
             {
-                throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
+                LogWritePermissionError();
+                return;
             }
 
             if (index < m_List.Length)
@@ -480,7 +484,8 @@ namespace Unity.Netcode
             // check write permissions
             if (!CanClientWrite(m_NetworkBehaviour.NetworkManager.LocalClientId))
             {
-                throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
+                LogWritePermissionError();
+                return;
             }
 
             var value = m_List[index];
@@ -505,7 +510,8 @@ namespace Unity.Netcode
                 // check write permissions
                 if (!CanClientWrite(m_NetworkBehaviour.NetworkManager.LocalClientId))
                 {
-                    throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
+                    LogWritePermissionError();
+                    return;
                 }
 
                 var previousValue = m_List[index];